Move extra_args to the end of the rubocop command
For some rubocop executables, a `--file` or `--path` option might be
necessary to target a specific file to run the parser.  This currently
would need to be the last argument in the string to be paired up with
the `l:filename` when running the RuboCop command.

Swapping these arguments should retain the previous functionality just
fine, while allowing for the use-case described above.

Note:  It was chosen to have `l:extra_args` last in the list versus
`a:current_args` since a `--file` option is more useful to be applied to
every command instead of on demand via the `a:current_args`.